Update SDL 1.2 backend to reflect recent changes
